Fixed a bug in SystemStorage and worked on code generation.

+++ blocks/src/hal.rs
@@ -5,4 +5,4 @@ use embedded_hal_0_2 as eh_02;
 
 pub type OutputPin<'a> = &'a mut dyn digital::OutputPin<Error = HalError>;
 
-pub type PwmPin<'a> = &'a mut dyn eh_02::PwmPin<Duty = f64>;
+pub type PwmPin<'a> = &'a mut dyn eh_02::PwmPin<Duty = u32>;
